requirements.txt
filterpy~=1.4.5
numpy~=1.20.0
Cython
watchdog
easydict==1.9
effdet==0.2.1
matplotlib==3.3.4
numpy==1.20.0
omegaconf==2.0.6
pandas==1.2.1
Pillow==8.1.0
pycocotools==2.0.2
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.0
sklearn
timm==0.3.4
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
xmltodict==0.12.0
opencv-python~=4.5.1.48
scikit-image~=0.18.1
requests
seaborn
torchaudio
environs~=9.3.1
grpcio-tools==1.33.2
grpcio==1.33.2
pytest